# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Cube Castle Internal API
  version: 1.2.0
  description: "元合约：定义城堡内部模块间的通信规约 - 业务ID系统"
paths:
...
  /api/v1/corehr/organizations:
    get:
      summary: List organizations
      operationId: listOrganizations
      tags:
        - CoreHR
      parameters:
        - name: include_uuid
          in: query
          description: Include UUID in response for system integration
          required: false
          schema:
            type: boolean
            default: false
        - name: search
          in: query
          description: Search by name or business ID
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationListResponse'
        '500':
          description: Internal server error

    post:
      summary: Create organization (auto-generates business ID)
      operationId: createOrganization
      tags:
        - CoreHR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
...
  /api/v1/corehr/organizations/{organization_id}:
    get:
      summary: Get organization by business ID
      operationId: getOrganization
      tags:
        - CoreHR
      parameters:
        - name: organization_id
          in: path
          required: true
          description: Organization business ID (100000-999999)
          schema:
            type: string
            pattern: '^[1-9][0-9]{5}$'
        - name: include_uuid
          in: query
          description: Include UUID in response for system integration
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Organization details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          description: Invalid business ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Organization not found
        '500':
          description: Internal server error

  # 业务ID管理 API
  /api/v1/corehr/business-ids/generate:
    post:
      summary: Generate business ID for entity (internal use)
      operationId: generateBusinessId
      tags:
        - CoreHR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateBusinessIdRequest'
      responses:
        '200':
          description: Business ID generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateBusinessIdResponse'
        '400':
          description: Invalid entity type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
  # 内部事件 API
...
components:
  schemas:
...
    Organization:
      type: object
      properties:
        # 业务ID (主要标识)
        id:
          type: string
          description: "Business ID - string format number (100000-999999)"
          pattern: '^[1-9][0-9]{5}$'
          example: "100000"
        # 系统UUID (仅当请求时包含)
        uuid:
          type: string
          format: uuid
          description: "System UUID - for internal integration"
          example: "550e8400-e29b-41d4-a716-446655440000"
        # 业务字段
        name:
          type: string
          example: "技术部"
        unit_type:
          type: string
          enum: [COMPANY, DEPARTMENT, TEAM]
          example: "DEPARTMENT"
        description:
          type: string
          nullable: true
          example: "负责技术开发和维护"
        level:
          type: integer
          description: "Organization hierarchy level"
          example: 1
        # 关联字段 (业务ID)
        parent_id:
          type: string
          description: "Parent organization business ID"
          pattern: '^[1-9][0-9]{5}$'
          nullable: true
          example: "100001"
        manager_id:
          type: string
          description: "Manager business ID"
          pattern: '^[1-9][0-9]{0,7}$'
          nullable: true
          example: "1"
        # 计算字段
        employee_count:
          type: integer
          description: "Number of employees in this organization"
          example: 15
        # 状态字段
        status:
          type: string
          enum: [ACTIVE, INACTIVE, PLANNED]
          default: ACTIVE
        # 系统字段
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - unit_type
        - level
    CreateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
          example: "技术部"
        unit_type:
          type: string
          enum: [COMPANY, DEPARTMENT, TEAM]
          example: "DEPARTMENT"
        description:
          type: string
          nullable: true
          example: "负责技术开发和维护"
        parent_id:
          type: string
          description: "Parent organization business ID"
          pattern: '^[1-9][0-9]{5}$'
          nullable: true
          example: "100001"
        manager_id:
          type: string
          description: "Manager business ID"
          pattern: '^[1-9][0-9]{0,7}$'
          nullable: true
          example: "1"
      required:
        - name
        - unit_type
...
    OrganizationListResponse:
      type: object
      properties:
        organizations:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        total_count:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: "Error type"
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: "Human-readable error message"
          example: "Invalid business ID format"
        details:
          type: object
          properties:
            field:
              type: string
              example: "employee_id"
            expected_format:
              type: string
              example: "1-99999999 (string format)"
            provided_value:
              type: string
              example: "invalid_id"
        validation_errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string

    # 业务ID管理模型
    GenerateBusinessIdRequest:
      type: object
      properties:
        entity_type:
          type: string
          enum: [employee, organization, position]
          description: "Type of entity to generate ID for"
          example: "employee"
        tenant_id:
          type: string
          format: uuid
          description: "Tenant ID for scoping"
          example: "00000000-0000-0000-0000-000000000000"
      required:
        - entity_type
        - tenant_id
    GenerateBusinessIdResponse:
      type: object
      properties:
        business_id:
          type: string
          description: "Generated business ID"
          example: "1"
        entity_type:
          type: string
          example: "employee"
        generated_at:
          type: string
          format: date-time
        id_ranges:
          type: object
          properties:
            employee:
              type: string
              example: "1-99999999"
            organization:
              type: string
              example: "100000-999999"
            position:
              type: string
              example: "1000000-9999999"
          description: "Available ID ranges for each entity type"